Refactored Video scenes.
